Various performance improvements related to album art generation in MediaProvider.
Avoid duplicating (or, in actuality 5-plicating) the amount of effort
we spend on creating album art thumbnails.

Process thumbnail creation requests in LIFO order rather than FIFO,
the assumption being that the user is looking at the items just added
to the queue and older ones may have already scrolled out of view.

Don't recompress thumbnail data if it's already of appropriate size,
instead, write the compressed data out as is.

related to bug 1742501