Implement MTP device to host notifications when files are added and removed.
Now MTP events are sent to the host when objects are added and deleted
due to device-side changes (rather than MTP initiated).

Also clean up MTP database upgrade path, fixing bug 2841422
This change is incompatible with the state of master over the last couple weeks
but will not be a problem for upgrades from earlier releases.

BUG: 2841422

Change-Id: I138d252e9b2b01371711eb7db723a908dc97e75b
Signed-off-by: Mike Lockwood <lockwood@android.com>