Use a database trigger to keep media tables in sync when deleting from files table
This is needed to allow deleting multiple files simultaneously.

Also move _OBJECT_REMOVED custom trigger to the new trigger so it will work for
non media files.

And while we are bumping the database version, also add a "title" column to the files
table, since we will be needing that soon.

Change-Id: I1d28cf9ba98c7c26c82738a908b7445e77b0bc85
Signed-off-by: Mike Lockwood <lockwood@android.com>