Delete database file and start from scratch if database upgrade fails.
This change is to handle a situation where we are unable to upgrade database
version due to lack of disk space or some other unusual condition.
If this occurs, then we simply delete the database file and try again.
This should result in an empty database file with the latest schema,
which will hopefully be repopulated when the media scanner runs at boot.

Change-Id: I0b00afc698a2d0761575d72223a6f8dd47ad4f70
Signed-off-by: Mike Lockwood <lockwood@android.com>