Optimize MediaThumbRequest RAM usage.
The MediaThumbRequest contains an BitmapFactory.Options instance
that is unused and a Random instance.

The BitmapFactory.Options instance can be removed and the Random
instance can be made static so that the object is shared among
the MediaThumbRequests. Since Random already is synchronized this
should mean no performance degrade and if the phone contains many
indexable objects the number of objects allocated goes down a lot.

Change-Id: I886948dc01baca9112cd345d939b63d3755ad235