Remove DATE_MODIFIED Reinsertion
Remove the lines where DATE_MODIFIED was retrieved from the map as
an int and being reinserted as an int. These lines weren't doing
anything besides getting the value and putting it back...

There is probably no impact, because the value for DATE_MODIFIED is
supposed to be in seconds. A CTS test that was using milliseconds
overflowed and caught this instance.

Change-Id: I185917d94805da98d1ff56fdcb2e4b0f5f857595