Fix MtpService connection/disconnection
In addition to fixing up synchronization, this avoids a problem where we
were trying to disconnect from MtpService when it was not connected.
This only occurred when USB was connected before the device was provisioned,
in which case the MediaProvider had not ever connected to MtpService
at the time USB was connected.

BUG: 3381229

Change-Id: I49e7090f5185ea6a158ae8f3e6656ae1767ff8fb
Signed-off-by: Mike Lockwood <lockwood@android.com>