Add support for multiple volumes to the media provider and MTP
In MTP each volume will appear as a separate storage unit.
However, in the provider, the database provides a unified view all the volumes.
We use the new storage_id column to tell which volume a given file is on.

Still to do: run the media scanner when volumes are added or removed.

Change-Id: I31f728beeb0f3b9a464343f0735b405b9f0ce79e
Signed-off-by: Mike Lockwood <lockwood@android.com>