DO NOT MERGE Add support for multiple volumes to the media provider and MTP
In MTP each volume will appear as a separate storage unit.
However, in the provider, the database provides a unified view all the volumes.
We use the new storage_id column to tell which volume a given file is on.

Still to do: run the media scanner when volumes are added or removed.

Change-Id: Ia6d1b610916438a1110b1900b66b7dca3d65d9a5
Signed-off-by: Mike Lockwood <lockwood@android.com>