Use argument binding in more places.
There were a number of places in the media provider where it would
construct the query by calling QueryBuilder.appendWhere() like so
for example: qb.appendWhere("_id = " + uri.getPathSegments().get(3));
This prevents effective use of the statement cache, because now
there is a compiled statement with a hardcoded id.
This change improves that by using argument binding, so it adds
"_id=?" to the where clause, and adds the value to substitute
to the selection args list.
b/3043992

Change-Id: I68641e59da9cd74f565b1d462a7dad44826785c8