Add parameter to allow speeding up deletes
In some cases we don't want the underlying file to be deleted when its
database entry is deleted. The media scanner does this by updating the
row and clearing its _data column before deleting it. This adds quite
a bit of overhead and slows down the scan when lots of entries are deleted.
This change adds support for a new parameter to be added to Uris when
deleting entries from the media provider, which makes the media provider
skip deleting the underlying file, which allows the media scanner to
delete entries from the database without first having to update them.

Change-Id: I9f83e6dd02d84166d9ee95cd466e763ab5624459