Remove more triggers
Remove the audio_delete and audio_meta_cleanup triggers and perform
their function in provider code.
Note that the audio_delete trigger was was incorrectly recreated in
the recreateAudioView() method, which made it so the trigger was
never actually used (since we're not deleting on the audio view anymore),
and even if you managed to use it somehow, it would have failed (because
one of its trigger actions was a delete on audio_meta, which is a view).

Change-Id: I846c3e24ae9b8ad539233f84714ef117e8ce6a9c