Fix media database after sdcard mountpoint move
The internal "sd card" storage is now in /storage/sdcard0. However
since /mnt/sdcard still exists for compatibility, the media scanner
will see a bunch of new files in /storage, in addition to the existing
files in /mnt/sdcard, resulting in two entries in the database for
every file.
This change removes any entries that were already scanned in /storage,
then updates the paths for the existing /mnt/sdcard entries to point
to /storage instead. This will handle both updates from ICS, as well
as fix the database for people that ran a build inbetween the move
to /storage and this change.

b/6131916

Change-Id: I1c071187651544be0fdb69e1cdb6367b636fc6d5