Don't create files when inserting entries
Don't automatically create files when inserting an image or video entry,
instead just generate the filename and store it in the database, and
defer creation of the file until the app calls openFile().
This could potentially break apps, but Camera and Gallery seem fine with
it, and CTS tests still pass.
b/7298303

Change-Id: Iaa311eecd6ae84f5748d8c7c3927a885f3b8c33f