Fix media provider deadlock
It was possible for the media provider to deadlock if writing album
art failed. In this case, it would call back in to the content provider
while inside a transaction. That new call would attempt to synchronize
on sGetTableAndWhereParam. If another thread happened to be doing an
update at the same time, then that thread could have already synchronized
on sGetTableAndWhereParam and be blocked on starting a transaction.
Moving the call outside of the transaction should fix this problem.
b/8114247

Change-Id: Ic8ca14b09de6221c4f99bfb323fd8c87c51c0225