Column in log table is not unique
DATETIME was set as primary key but that value is not always unique.
This solution is to remove primary key constraint from column "time"
to avoid SQLiteConstraintException caused in case same DATETIME value
is attempted to be inserted into log table.

(this is externally contributed change ef8c5a3947f592154e443e7c63c3e6f2ebde8a07
which was skipped in an earlier automerge)
b/8832525

Change-Id: I26ba5356f54b589882447f4aa51892e6b2928acd