Update primary storage paths when volumes change.
Devices that use physical media as their primary storage may not
have that storage mounted by the time that MediaProvider is started,
which means we'd snapshot a stale primary storage path.

To fix this, refresh those paths as volumes are mounted or unmounted
over time.  Also fix logic that detects secondary storage devices.

Bug: 22629557
Change-Id: I3c6b057ce082318e1b28e261c953a436da42e36f