Fix MediaProvider to report details of paths that change.
If you want to observe media changes with the job scheduler, we'd
like to give you details of the files that changed when possible.
However, MediaProvider was only reporting the containing path of
a change, not the actual URI that changed.

Fix this by now reporting the newUri when doing an insert.

There is a comment in there about whether we should continue also
reporting the parent URI.  I am not sure what to do about that.

Change-Id: I03667c4466bde933cf2e04f8aca85eac913609f7