Be smarter about reporting changes to containing URIs.
This is for work on issue #26390151 (Add new JobScheduler API...
for monitoring content providers)

The media provider uses the new content observer flag when
reporting a change to the containing URI of a specific file
changing, so content observers can avoid seeing that if they
are already listening for the specific file.  This allows
observers to avoid getting dirty data that prevents them
from telling that only a specific URI has changed.

Change-Id: I38b4c1941f1b06c121a1caf2b5d8593da2007410