Stop instantiating second MtpServer object.
MtpServer thread opens MTP device file, and only one thread can open the
device file at the same time.

Previously Android tried to instantiate multiple MtpServer instances and
the second instance of MtpServer failed to open MTP device. It happens
when MtpService which holds MtpServer is destroyed and created again
in short time.

Not to create the second instance of MtpServer, the CL introduces new
class ServerHolder and keeps it in static field of
MtpService. ServerHolder acts as a cache for MtpServer and prevent
MtpServer from being instantiated multiple times.

BUG=29971895
TEST=Connect Android to workstation and use MTP/PTP mode.

Change-Id: I9f8f1304a6c5a2b38d6b6abde6bdf1f0992feefd