[RingtonePicker] Get cursor to list parent sounds
Since it is now possible for the work profile to use ringtones from the
personal profile, even custom ringtones, then it makes sense for the
RingtonePicker to list them as well.

This changes use a new RingtoneManager API to also include the personal
ringtones in the work profile RingtonePicker, also displaying a small
work badge next to the custom ringtones of the work profile.

Test: manual - install custom ringtone in work profile and see if it shows with a work badge in the picker
Bug: 30658854
Change-Id: I76e46e7d9401d54e8f096b80ea79e643ce2ab37e