Also configure if enabled and connected but not configured.
We don't know if the media process has been killed before this
usb connection has been started. If it was killed, then the
device will connect but not be configured. Adding a configure call
on connect will ensure that the device is configured later.

If the media process has not been killed, then the configure will
do nothing since the previous configure will have already done the
work.

Bug: 36801389
Test: Run am kill-all, verify media process died, plug in and verify usb
still works.

Change-Id: I6d0b1854f3bf8973d27d6ca22ce438bcbf6982af