Configure Mtp once on boot.
It is possible that UsbDeviceManager ends up
sending USB_STATE intents during a time when
MtpReceiver cannot start to process them.
In these cases, mtp will never be configured.

To fix, configure once on boot regardless. This
will be harmless if Mtp is already configured
since the additional configure will do nothing,
or if Mtp is not enabled.

Bug: 64197549
Test: set require pass on boot -> reboot -> unlock -> plug in -> verify
Test: unset require pass on boot -> reboot -> unlock -> plug in ->
verify
Test: above two tests with usb debugging on/off
Change-Id: I5a96078c2ebb6517e226adedd0f0ca8fb09c15c2