Make MtpReceiver and MtpService single user.
Mtp uses file descriptors that only one process
can hold at a time. Instead of starting a different
MtpService for each user and quitting if it is
not the foreground user, have one centralized MtpService
that only exposes the volumes of the current foreground
user.

MtpService will use ActivityManager.getCurrentUser()
to determine which user's data is displayed.

Bug: 64822515
Test: Use MTP and PTP, switch users, verify correct storage is shown
Change-Id: I8dfedb812f411833075772bd5fd76ce042ec14dc