Don't leak MtpDatabase receiver.
There is a race on usb disconnect between
MtpServer and UsbDeviceManager. If the
kernel disconnect signal is sent to MtpService
before MtpServer, the service will be stopped
before the receiver is unregistered.

To fix this, set the server in the database to
null when service stops, which also unregisters
the receiver. Addtional guards in MtpDatabase
make it robust to a null server.

Bug: 76433619
Test: no error
Change-Id: I99dab3d2fb5884262b680e6a3e5eb4c5fb6a3457