Only set RESULT_CANCELED for RingtonePickerActivity if it's explicitly canceled.
In handset, a RESULT_OK status with a null EXTRA_RINGTONE_PICKED_URI indicates
an affirmative selection of the "Silent" ringtone. We were previously returning
RESULT_CANCELED if the selection was the same as the previously-selected
ringtone (which would over-trigger when clients omitted the
EXTRA_RINGTONE_EXISTING_URI, indicating that they don't know any existing
ringtone, but not necessarily that they would default to 'silent' if the
picker is canceled). After this change, we'll correctly return RESULT_OK
with a null extra.

In Wear, no client requires us to set RESULT_CANCELED just to indicate that the
selection was unchanged or restored to default by the time the user left the
activity, and since we have no explicit "cancel" button the result is just the
same as if the user re-selected the same ringtone; we can simply use RESULT_OK
every time.

Bug: 67351809
Fixes: 67351809
Test: manual - repro'd with the RingtonePicker.apk from b/67351809 before making
      this change, and then observed the correct behavior after the change.

Change-Id: I7601618d14af020a8b135a36838fcea0f087ebb9