Support RRO for ringtones, notifications and alarms
Copy ringtone, alarm and notification from resources if an RRO has been
applied to the com.android.providers.media package. Files are copied to
the standard directories for each category and if their metadata points
to string resources, the overlayed string resources will be used by
MediaProvider including when language is changed on the device.

Note that the default resources are empty binaries and a default sound
overlay will only be copied if its size is not zero.

Bug: 69299998
Test: Verify that a newly installed and enabled test RRO with one or
      more sound resources changes the corresponding default sounds.
Change-Id: I3bba5ad9e59c1f9aeb4af7a9266f073c49c2df74