Move updates and deletes to SQLiteQueryBuilder.
As part of the new storage work in Q, we're going to be relying
much more on MediaProvider to enforce security boundaries, so we
need to tighten things up.

This change starts using a newly added SQLiteQueryBuilder feature
that supports update() and delete() calls, and a future change
will enable "strict" enforcement to combat SQL injection attacks.

It also uses a new appendWhere() feature to simplify selection
argument tracking.

Bug: 111268862
Test: atest cts/tests/tests/provider/src/android/provider/cts/MediaStore*
Change-Id: Ida08b81b5472811bd04b00672d7a3677253eba93