Start merging query and update/delete builders.
This has multiple benefits: first, it makes it easier to ensure that
all code paths are building WHERE clauses the same way, meaning less
risk of subtle bugs.  And second, it'll make it easier to enforce the
upcoming Q security model in one code path.

This pass is doing the simple query() migrations that are mostly
mechanical in nature; future CLs will finish the remaining queries.

Remove hidden APIs related to search and MTP, neither of which are
actively being used.

Bug: 111268862
Test: atest cts/tests/tests/provider/src/android/provider/cts/MediaStore*
Change-Id: If7d809d4388bbb647bd004334ea9df4afe9edcc5