Band-aid for apps relying on implementation.
We found out that a specific Camera app is inserting an item into
the Files table, but then trying to modify it via the Images table,
which should have failed, but the older implementation details
let it pass through.

To unblock the chaselist, provide narrow legacy behavior to work
around the app bug without requiring them to re-spin.

Bug: 111966296
Test: manual
Change-Id: I3977ebb9fd125889fec39e7e05ea657d03fa8a9b