Another band-aid for apps relying on implementation.
We found out that a specific Camera app is inserting an item into
the Files table, but then trying to modify it via the Video table,
which should have failed, but the older implementation details
let it pass through.

To unblock the chaselist, provide narrow legacy behavior to work
around the app bug without requiring them to re-spin.

Bug: 112246630
Test: manual
Change-Id: Ic98a605e43f7f2a7ec2bf582d727dbf62f942505