Add "owner_package_name" column.
To support new storage features in Q, we need to start tracking the
"owner" of each inserted file.  This change introduces the schema
change to insert the column, including logic to derive values for
existing well-known package-specific paths.  Populate the column for
any newly inserted media based on the calling package.

Tests to verify upgrade and downgrade behavior.

Bug: 111893193
Test: adb shell am instrument -w com.android.providers.media.tests
Change-Id: I34c7d10ca150622dc7f92f658014fee2c3df3e8b