Orphan media when app uninstalled or data cleared.
When an app contributes media, we're tracking them as the "owner"
of that media.  When the user uninstalls that app, or clear data on
that app, content that they've contributed is retained for the user,
but it's now "orphaned" and no longer associated with the app that
created it.

This also creates a nightly idle maintenance service that does
sanity checking to handle cases where a broadcast was dropped due
to crash or rapid restart.

Bug: 111893193
Test: manual
Change-Id: I772c553df8a4d308061ff156ca67677d9957ccfd