Define app-ops for new storage permissions in Q.
Mostly mirrored off existing READ/WRITE_EXTERNAL_STORAGE app-ops,
but with new Aural/Visual runtime permission groups called out.

Starts noting these ops in MediaProvider when callers are accessing
data which is protected by the relevant permissions.  Simplifies
the logic used to check for permissions, so we only check them in
the context of a specific media type.

If the caller is part of the media system, or if they hold a Uri
grant, then we don't need to note the operation.  (The operations
are tracking direct usage of the runtime permission.)

Bug: 117777783
Test: atest cts/tests/tests/provider/src/android/provider/cts/MediaStore*
Change-Id: If3982f7fb1c124d16832736df04c76dddfbffc94