Trigger scan when apps attempt to mutate metadata.
Some apps are modifying the data directly on disk, instead of going
through ContentResolver.openFile(), which means we can't detect
their changes.  Thankfully those apps typically then update() the
items, which is a good secondary signal that we should trigger a
re-scan of the item.

We only re-scan after applying the update, since the caller
have just manually moved the file.

Bug: 120711487
Test: manual
Change-Id: I8a1421d7d2ecdc77de09333279a78789a0516cab