Use new @Column annotation for all projections.
A recent change defined a new annotation that "contract" classes
can use to clearly describe the columns that are available for
insert(), update(), and query() when working with ContentProviders.

This change uses those annotated columns to build the list of
allowed columns for each incoming query.  This uncovered several
obscure bugs where the provider wasn't technically meeting the
public API contract, and those are now fixed so we fully meet the
API contracts.

Bug: 120429729
Test: atest cts/tests/tests/provider/src/android/provider/cts/MediaStore*
Change-Id: Iab8716b08f80c275f7f171d25afcf018f063e481