Forget metadata scanned from stale volumes.
We're willing to keep around scanned metadata for one week, to help
speed up cases where the user ejects and reinserts the same storage
device over time.

But we draw the line at one week as a reasonable trade-off: if we
haven't seen a device in that long, then forget any scanned
metadata.  We can easily re-scan that data if the user inserts the
device again.

Bug: 131751531
Test: atest --test-mapping packages/providers/MediaProvider
Change-Id: I69881922cc434897f3e899702c5de217adabd5a5