Selectively redact XMP data
Instead of redacting the whole XMP block, redact only xmp tags
known to contain GPS locations. For now, this is limited to the
GPS tags in the exif namespace.

This also fixes a bug where XMP was not redacted at all for
media where the XMP was found via its guid in IsoInterface.

Since the XmlPullParser only tells you the line/column offsets
of the data it's reading, the input stream it's reading from
had to be amended to map newline->byte_offset.

Bug: 132818091
Test: two new CTS tests added
Test: atest --test-mapping packages/providers/MediaProvider
Change-Id: I4400654e32505d0d3882bb708c234a9f352b0eab