Set notification Uri, but don't register.
Several clients like MediaDocumentsProvider copy these notification
Uris for their external clients to monitor, so we need to keep
populating them.  However, what we're really trying to avoid is the
overhead of registering an in-process ContentObserver, since we
don't display any UI lists outselves.

This CL offers that middle ground: we always set the notification
Uris, but without registering ourselves for changes.

Bug: 133124199
Test: atest --test-mapping packages/providers/MediaProvider
Change-Id: I76d79d29b249cc8dce8b3cc632cb04c342e4a6a1