MediaProvider: Reorder object removed callback to fix internal.db issue
In a previous CL, mInternal for internal.db was flipped to false,
allowing extra tables/objects to be created. This allowed the
possibility of calling _OBJECT_REMOVED(), which is supposed to be
external only.

A later CL added a misguided bug fix that re-flipped flag back to
true, causing the method to not be added for internal.db. Rather
than re-flip the flag again and going back into the rabbit hole of
broken upgrades, this adds the method regardless of the flag.

Bug: 134943032

Test: ran MediaStore_FilesTest

Change-Id: I8f81d4d92772740f3b84f90d9f5f67f8550be801