Introduce device sys prop for toggling redaction
Add check to FuseDaemon on reads to see whether to redact or not.
The check is controlled by the newly added system property.
The system property is checked on the first read of every file, meaning
that it's hot swappable.

The default behaviour is equivalent to the property being set to true.
Sys prop name: persist.sys.fuse.redaction-enabled

Test: adb pull <photo-with-location-meta-data> # with the property on,
off and unset
Fixes: 142240126

Change-Id: Ibb39005ab9a41606312e98370be228820825dd15