Wire file creation to go through MediaProvider
Make FuseDaemon call MediaProvider API to create a file on the app's
behalf instead of going directly to the lower FS.

MediaProvider ensures we track file ownership, type-path conformity and
that no app can create a file in other apps' specific directories.
In case an app tries to create a file in other app's specific directory,
ENOENT will be returned regardless of whether the other directory exists
or not (see: b/113074599).

Test: temporary tests in experimental dir
Test: b/142926859 - will add CTS tests soon
Bug: 142476362
Bug: 142475473
Change-Id: Iaab206735a71765b9b55c1268d4431dc8e1beed1