Make the android.process.media a persistent process
Now the FUSE daemon runs in the media process, lots of critical
processes now depend on it. For instance, the sysUI and even vold.

If the FUSE daemom crashes, the ActivityManager service will throttle
it's restart, further putting the entire system in a bad state since
critical processes will now block on any IO on external storage.

Test: m
Bug: 135341433
Change-Id: I910e0e2ed7d3289f64a75b6e5bf5ca056722c2eb