Progress towards splitting MediaProvider UID.
We're working towards making MediaProvider into a Mainline module,
which means it can no longer share the "android.media" UID built
into the OS.  The major reason for this is because they're signed
with separate keys, but a secondary reason is that we don't know what
database customizations an OEM has performed in their MediaProvider.

This initial pass introduces the APIs needed to support a "legacy"
MediaProvider that will be used to migrate data between the old and
new UID.  This legacy provider is implemented using the same code
as the modern provider, but hosted at the "media_legacy" authority.

Bug: 144247087
Test: atest --test-mapping packages/providers/MediaProvider
Change-Id: I223dc16837f3655d71437e4e852c024a7013e9c0