Progress towards splitting MediaProvider UID.
We're working towards making MediaProvider into a Mainline module,
which means it can no longer share the "android.media" UID built
into the OS.  The major reason for this is because they're signed
with separate keys, but a secondary reason is that we don't know what
database customizations an OEM has performed in their MediaProvider.

Move to using JobIntentService, since MediaReceiver won't be able to
start services from the background.

Bug: 144247087
Test: atest --test-mapping packages/providers/MediaProvider
Change-Id: Ia5db9d27659177f4dd14a787e89213830a3a2680