Don't fail getRedactionRanges on FileNotFoundException
This is a workaround for cases where app A creates a db entry but the
actual content is filled by app B. In this case, the file is not created
during insertion, but it's created when it's opened for the first time
(for write). Usually, the owner of the file bypasses the redacion
logic, but since the file is created by app B, it won't exist by the
time of opening and we need to ignore FileNotFoundException.
This shouldn't be a security/privacy issue since the file is not yet
created and so it makes sense that there's nothing to redact.

This bug was just discovered due to recent refactoring in the
redaction logic: before the refactoring, we ignored all exceptions
thrown during calculating the redaction ranges.

Test: atest --test-mapping
Test: manual - take snapshot from Facebook Lite app
Bug: 144752607
Change-Id: I1c3c737e5e3d35ac7ac793d389ff0b577ed03310